Code cleanup - move logic to extensions
Code cleanup - move logic to extensions
