+++ android/src/main/java/com/appfolio/work/UploadWorker.kt
@@ -3,26 +3,18 @@ package com.appfolio.work
 import android.content.Context
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.google.gson.Gson
-import net.gotev.uploadservice.data.UploadNotificationConfig
-import net.gotev.uploadservice.data.UploadTaskParameters
-import net.gotev.uploadservice.persistence.PersistableData
+import com.appfolio.extensions.PARAM_KEY_NOTIF_CONFIG
+import com.appfolio.extensions.PARAM_KEY_TASK_PARAMS
+import com.appfolio.extensions.toUploadNotificationConfig
+import com.appfolio.extensions.toUploadTaskParameters
 
 class UploadWorker(val context: Context, params: WorkerParameters): Worker(context, params) {
 
-  companion object {
-    const val PARAM_KEY_TASK_PARAMS = "TASK_PARAMS"
-    const val PARAM_KEY_NOTIF_CONFIG = "NOTIF_CONFIG"
-  }
-
   override fun doWork(): Result {
-    val gson = Gson()
-    val taskParamsStr = inputData.getString(PARAM_KEY_TASK_PARAMS)
-    val uploadTaskParameters= UploadTaskParameters.createFromPersistableData(PersistableData.fromJson(taskParamsStr!!))
-    val configStr = inputData.getString(PARAM_KEY_NOTIF_CONFIG)
-    val notificationConfig = gson.fromJson(configStr, UploadNotificationConfig::class.java)
+    val taskParamsStr = inputData.getString(PARAM_KEY_TASK_PARAMS) ?: return Result.failure()
+    val notifConfigStr = inputData.getString(PARAM_KEY_NOTIF_CONFIG) ?: return Result.failure()
 
-    UploadManager.startUpload(context, uploadTaskParameters, notificationConfig)
+    UploadManager.startUpload(context, taskParamsStr.toUploadTaskParameters(), notifConfigStr.toUploadNotificationConfig())
     return Result.success()
   }
 }
